+++ src/run/ui_main.py
@@ -5,7 +5,7 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from PyQt5.QtWidgets import QApplication
 from loguru import logger
-from ui_event.base_event import exit_save_data
+from ui_event.base_event import exit_save_data, kill_other_close
 from log.log_record import log_record, check_version
 from ui_event.pyqt_main_ui import UIMainWindows
 from utils.sytem_monitor import sys_mon
@@ -36,6 +36,8 @@ def on_last_window_closed():
     """
     logger.warning("Console window is closing...")
     exit_save_data()
+    logger.warning("start clean other process...")
+    kill_other_close()
 
 
 @logger.catch